+++ R/fstadv.R
@@ -8,11 +8,8 @@
 #' @export
 get_fstadv <- function(link) {
 
-    # Check status of link(s)
-    valid.link <- purrr::map_chr(link, status) %>% stats::na.omit()
-
     # Get all products for the current storm
-    products <- purrr::map(valid.link, get_products) %>% purrr::flatten_chr()
+    products <- purrr::map(link, get_products) %>% purrr::flatten_chr()
 
     # Filter out fstadv products
     products <- filter_fstadv(products)