+++ tests/testthat/test_discus.R
@@ -1,5 +1,11 @@
 context("Storm Discussions (discus)")
 
+# Set timeout options
+opt.timeout <- getOption("rrricanes.http_timeout")
+opt.attempts <- getOption("rrricanes.http_attempts")
+options("rrricanes.http_timeout" = 1)
+options("rrricanes.http_attempts" = 5)
+
 ## ---- 2008, AL ---------------------------------------------------------------
 al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
 
@@ -26,3 +32,7 @@ test_that("Dataframe Skeleton", {
 test_that("Test get_discus()", {
     expect_identical(al092008.discus, df.al092008.discus)
 })
+
+# Reset options
+options("rrricanes.http_timeout" = opt.timeout)
+options("rrricanes.http_attempts" = opt.attempts)