+++ tests/testthat/test_public.R
@@ -1,5 +1,11 @@
 context("Public Advisory (public)")
 
+# Set timeout options
+opt.timeout <- getOption("rrricanes.http_timeout")
+opt.attempts <- getOption("rrricanes.http_attempts")
+options("rrricanes.http_timeout" = 1)
+options("rrricanes.http_attempts" = 5)
+
 ## ---- 2008, AL ---------------------------------------------------------------
 al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
 
@@ -12,3 +18,7 @@ load(system.file("extdata", "al092008.public.Rda", package = "rrricanes"))
 test_that("Test get_public()", {
     expect_equal(al092008.public, df.al092008.public)
 })
+
+# Reset options
+options("rrricanes.http_timeout" = opt.timeout)
+options("rrricanes.http_attempts" = opt.attempts)